Separate out CarouselIndicator and other cleanup
