+++ src/nft/components/explore/CarouselIndicator.css.ts
@@ -1,0 +1,9 @@
+import { sprinkles } from 'nft/css/sprinkles.css'
+
+export const carouselIndicator = sprinkles({
+  width: '36',
+  height: '4',
+  marginRight: '6',
+  borderRadius: 'round',
+  display: 'inline-block',
+})
